refactor: move bundle to Localizer, change codestyle
