+++ robots/src/main/java/game/views/locale/MenuItem.java
@@ -5,19 +5,16 @@ import java.util.ResourceBundle;
 
 public class MenuItem implements LocaleChangeListener {
     private final JMenuItem item;
-    public JMenuItem getItem(){ return this.item; }
     private final String label;
+    public MenuItem(JMenuItem item, String label){
+        this.item = item;
+        this.label = label;
+    }
     @Override
     public void onLocaleChanged(ResourceBundle bundle) {
         item.setText(bundle.getString(label));
     }
-    private MenuItem(JMenuItem item, String label){
-        this.item = item;
-        this.label = label;
-    }
-    public static MenuItem instantiate(JMenuItem item, String label){
-        return new MenuItem(item, label);
-    }
+    public JMenuItem getItem(){ return this.item; }
     public void add(MenuItem item){
         this.item.add(item.getItem());
     }
